requirements.txt
ultralytics>=8.0.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
requests>=2.31.0
